Move completion implementation to completion/regexp_completor for future extension
