+++ test/irb/test_completion.rb
@@ -1,67 +1,88 @@
 # frozen_string_literal: false
 require "pathname"
 require "irb"
+require "irb/completion/regexp_completor"
 
 require_relative "helper"
 
 module TestIRB
   class CompletionTest < TestCase
+    DummyContext = Struct.new(:workspace)
+    DummyWorkSpace = Struct.new(:binding)
+
     def setup
       # make sure require completion candidates are not cached
       IRB::InputCompletor.class_variable_set(:@@files_from_load_path, nil)
+
+      # IRB::InputCompletor::CompletionProc requires IRB.conf[:MAIN_CONTEXT] to be present
+      IRB.conf[:MAIN_CONTEXT] = DummyContext.new(DummyWorkSpace.new(binding))
+    end
+
+    def teardown
+      IRB.conf[:MAIN_CONTEXT] = nil
+    end
+
+    def completion_candidates(target, bind)
+      completor = IRB::InputCompletor::RegexpCompletor.new(target, nil, nil, bind: bind)
+      completor.completion_candidates
+    end
+
+    def doc_namespace(target, bind)
+      completor = IRB::InputCompletor::RegexpCompletor.new(target, nil, nil, bind: bind)
+      completor.doc_namespace(target)
     end
 
     class MethodCompletionTest < CompletionTest
       def test_complete_string
-        assert_include(IRB::InputCompletor.retrieve_completion_data("'foo'.up", bind: binding), "'foo'.upcase")
+        assert_include(completion_candidates("'foo'.up", binding), "'foo'.upcase")
         # completing 'foo bar'.up
-        assert_include(IRB::InputCompletor.retrieve_completion_data("bar'.up", bind: binding), "bar'.upcase")
-        assert_equal("String.upcase", IRB::InputCompletor.retrieve_completion_data("'foo'.upcase", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("bar'.up", binding), "bar'.upcase")
+        assert_equal("String.upcase", doc_namespace("'foo'.upcase", binding))
       end
 
       def test_complete_regexp
-        assert_include(IRB::InputCompletor.retrieve_completion_data("/foo/.ma", bind: binding), "/foo/.match")
+        assert_include(completion_candidates("/foo/.ma", binding), "/foo/.match")
         # completing /foo bar/.ma
-        assert_include(IRB::InputCompletor.retrieve_completion_data("bar/.ma", bind: binding), "bar/.match")
-        assert_equal("Regexp.match", IRB::InputCompletor.retrieve_completion_data("/foo/.match", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("bar/.ma", binding), "bar/.match")
+        assert_equal("Regexp.match", doc_namespace("/foo/.match", binding))
       end
 
       def test_complete_array
-        assert_include(IRB::InputCompletor.retrieve_completion_data("[].an", bind: binding), "[].any?")
-        assert_equal("Array.any?", IRB::InputCompletor.retrieve_completion_data("[].any?", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("[].an", binding), "[].any?")
+        assert_equal("Array.any?", doc_namespace("[].any?", binding))
       end
 
       def test_complete_hash_and_proc
         # hash
-        assert_include(IRB::InputCompletor.retrieve_completion_data("{}.an", bind: binding), "{}.any?")
-        assert_equal(["Proc.any?", "Hash.any?"], IRB::InputCompletor.retrieve_completion_data("{}.any?", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("{}.an", binding), "{}.any?")
+        assert_equal(["Proc.any?", "Hash.any?"], doc_namespace("{}.any?", binding))
 
         # proc
-        assert_include(IRB::InputCompletor.retrieve_completion_data("{}.bin", bind: binding), "{}.binding")
-        assert_equal(["Proc.binding", "Hash.binding"], IRB::InputCompletor.retrieve_completion_data("{}.binding", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("{}.bin", binding), "{}.binding")
+        assert_equal(["Proc.binding", "Hash.binding"], doc_namespace("{}.binding", binding))
       end
 
       def test_complete_numeric
-        assert_include(IRB::InputCompletor.retrieve_completion_data("1.positi", bind: binding), "1.positive?")
-        assert_equal("Integer.positive?", IRB::InputCompletor.retrieve_completion_data("1.positive?", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("1.positi", binding), "1.positive?")
+        assert_equal("Integer.positive?", doc_namespace("1.positive?", binding))
 
-        assert_include(IRB::InputCompletor.retrieve_completion_data("1r.positi", bind: binding), "1r.positive?")
-        assert_equal("Rational.positive?", IRB::InputCompletor.retrieve_completion_data("1r.positive?", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("1r.positi", binding), "1r.positive?")
+        assert_equal("Rational.positive?", doc_namespace("1r.positive?", binding))
 
-        assert_include(IRB::InputCompletor.retrieve_completion_data("0xFFFF.positi", bind: binding), "0xFFFF.positive?")
-        assert_equal("Integer.positive?", IRB::InputCompletor.retrieve_completion_data("0xFFFF.positive?", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("0xFFFF.positi", binding), "0xFFFF.positive?")
+        assert_equal("Integer.positive?", doc_namespace("0xFFFF.positive?", binding))
 
-        assert_empty(IRB::InputCompletor.retrieve_completion_data("1i.positi", bind: binding))
+        assert_empty(completion_candidates("1i.positi", binding))
       end
 
       def test_complete_symbol
-        assert_include(IRB::InputCompletor.retrieve_completion_data(":foo.to_p", bind: binding), ":foo.to_proc")
-        assert_equal("Symbol.to_proc", IRB::InputCompletor.retrieve_completion_data(":foo.to_proc", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates(":foo.to_p", binding), ":foo.to_proc")
+        assert_equal("Symbol.to_proc", doc_namespace(":foo.to_proc", binding))
       end
 
       def test_complete_class
-        assert_include(IRB::InputCompletor.retrieve_completion_data("String.ne", bind: binding), "String.new")
-        assert_equal("String.new", IRB::InputCompletor.retrieve_completion_data("String.new", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("String.ne", binding), "String.new")
+        assert_equal("String.new", doc_namespace("String.new", binding))
       end
     end
 
@@ -160,13 +181,13 @@ module TestIRB
         local_variables.clear
         instance_variables.clear
 
-        assert_include(IRB::InputCompletor.retrieve_completion_data("str_examp", bind: binding), "str_example")
-        assert_equal("String", IRB::InputCompletor.retrieve_completion_data("str_example", bind: binding, doc_namespace: true))
-        assert_equal("String.to_s", IRB::InputCompletor.retrieve_completion_data("str_example.to_s", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("str_examp", binding), "str_example")
+        assert_equal("String", doc_namespace("str_example", binding))
+        assert_equal("String.to_s", doc_namespace("str_example.to_s", binding))
 
-        assert_include(IRB::InputCompletor.retrieve_completion_data("@str_examp", bind: binding), "@str_example")
-        assert_equal("String", IRB::InputCompletor.retrieve_completion_data("@str_example", bind: binding, doc_namespace: true))
-        assert_equal("String.to_s", IRB::InputCompletor.retrieve_completion_data("@str_example.to_s", bind: binding, doc_namespace: true))
+        assert_include(completion_candidates("@str_examp", binding), "@str_example")
+        assert_equal("String", doc_namespace("@str_example", binding))
+        assert_equal("String.to_s", doc_namespace("@str_example.to_s", binding))
       end
 
       def test_complete_sort_variables
@@ -176,7 +197,7 @@ module TestIRB
         xzy_1.clear
         xzy2.clear
 
-        candidates = IRB::InputCompletor.retrieve_completion_data("xz", bind: binding, doc_namespace: false)
+        candidates = completion_candidates("xz", binding)
         assert_equal(%w[xzy xzy2 xzy_1], candidates)
       end
     end
@@ -189,12 +210,12 @@ module TestIRB
       end
 
       def test_complete_constants
-        assert_equal(["Foo"], IRB::InputCompletor.retrieve_completion_data("Fo", bind: binding))
-        assert_equal(["Foo::B1", "Foo::B2", "Foo::B3"], IRB::InputCompletor.retrieve_completion_data("Foo::B", bind: binding))
-        assert_equal(["Foo::B1.positive?"], IRB::InputCompletor.retrieve_completion_data("Foo::B1.pos", bind: binding))
+        assert_equal(["Foo"], completion_candidates("Fo", binding))
+        assert_equal(["Foo::B1", "Foo::B2", "Foo::B3"], completion_candidates("Foo::B", binding))
+        assert_equal(["Foo::B1.positive?"], completion_candidates("Foo::B1.pos", binding))
 
-        assert_equal(["::Forwardable"], IRB::InputCompletor.retrieve_completion_data("::Fo", bind: binding))
-        assert_equal("Forwardable", IRB::InputCompletor.retrieve_completion_data("::Forwardable", bind: binding, doc_namespace: true))
+        assert_equal(["::Forwardable"], completion_candidates("::Fo", binding))
+        assert_equal("Forwardable", doc_namespace("::Forwardable", binding))
       end
     end
 
@@ -203,13 +224,13 @@ module TestIRB
         # trigger PerfectMatchedProc to set up RDocRIDriver constant
         IRB::InputCompletor::PerfectMatchedProc.("foo", bind: binding)
 
-        @original_use_stdout = IRB::InputCompletor::RDocRIDriver.use_stdout
+        @original_use_stdout = IRB::InputCompletor::RegexpCompletor::RDocRIDriver.use_stdout
         # force the driver to use stdout so it doesn't start a pager and interrupt tests
-        IRB::InputCompletor::RDocRIDriver.use_stdout = true
+        IRB::InputCompletor::RegexpCompletor::RDocRIDriver.use_stdout = true
       end
 
       def teardown
-        IRB::InputCompletor::RDocRIDriver.use_stdout = @original_use_stdout
+        IRB::InputCompletor::RegexpCompletor::RDocRIDriver.use_stdout = @original_use_stdout
       end
 
       def test_perfectly_matched_namespace_triggers_document_display
@@ -294,34 +315,34 @@ module TestIRB
       rescue
       end
       symbols += [:aiueo, :"aiu eo"]
-      candidates = IRB::InputCompletor.retrieve_completion_data(":a", bind: binding)
+      candidates = completion_candidates(":a", binding)
       assert_include(candidates, ":aiueo")
       assert_not_include(candidates, ":aiu eo")
-      assert_empty(IRB::InputCompletor.retrieve_completion_data(":irb_unknown_symbol_abcdefg", bind: binding))
+      assert_empty(completion_candidates(":irb_unknown_symbol_abcdefg", binding))
       # Do not complete empty symbol for performance reason
-      assert_empty(IRB::InputCompletor.retrieve_completion_data(":", bind: binding))
+      assert_empty(completion_candidates(":", binding))
     end
 
     def test_complete_invalid_three_colons
-      assert_empty(IRB::InputCompletor.retrieve_completion_data(":::A", bind: binding))
-      assert_empty(IRB::InputCompletor.retrieve_completion_data(":::", bind: binding))
+      assert_empty(completion_candidates(":::A", binding))
+      assert_empty(completion_candidates(":::", binding))
     end
 
     def test_complete_absolute_constants_with_special_characters
-      assert_empty(IRB::InputCompletor.retrieve_completion_data("::A:", bind: binding))
-      assert_empty(IRB::InputCompletor.retrieve_completion_data("::A.", bind: binding))
-      assert_empty(IRB::InputCompletor.retrieve_completion_data("::A(", bind: binding))
-      assert_empty(IRB::InputCompletor.retrieve_completion_data("::A)", bind: binding))
-      assert_empty(IRB::InputCompletor.retrieve_completion_data("::A[", bind: binding))
+      assert_empty(completion_candidates("::A:", binding))
+      assert_empty(completion_candidates("::A.", binding))
+      assert_empty(completion_candidates("::A(", binding))
+      assert_empty(completion_candidates("::A)", binding))
+      assert_empty(completion_candidates("::A[", binding))
     end
 
     def test_complete_reserved_words
-      candidates = IRB::InputCompletor.retrieve_completion_data("de", bind: binding)
+      candidates = completion_candidates("de", binding)
       %w[def defined?].each do |word|
         assert_include candidates, word
       end
 
-      candidates = IRB::InputCompletor.retrieve_completion_data("__", bind: binding)
+      candidates = completion_candidates("__", binding)
       %w[__ENCODING__ __LINE__ __FILE__].each do |word|
         assert_include candidates, word
       end
@@ -342,11 +363,11 @@ module TestIRB
       }
       bind = obj.instance_exec { binding }
 
-      assert_include(IRB::InputCompletor.retrieve_completion_data("public_hog", bind: bind), "public_hoge")
-      assert_include(IRB::InputCompletor.retrieve_completion_data("public_hoge", bind: bind, doc_namespace: true), "public_hoge")
+      assert_include(completion_candidates("public_hog", bind), "public_hoge")
+      assert_include(doc_namespace("public_hoge", bind), "public_hoge")
 
-      assert_include(IRB::InputCompletor.retrieve_completion_data("private_hog", bind: bind), "private_hoge")
-      assert_include(IRB::InputCompletor.retrieve_completion_data("private_hoge", bind: bind, doc_namespace: true), "private_hoge")
+      assert_include(completion_candidates("private_hog", bind), "private_hoge")
+      assert_include(doc_namespace("private_hoge", bind), "private_hoge")
     end
   end
 end
